+++ Code/LYRUIMessageCollectionViewCell.m
@@ -29,7 +29,7 @@
         [self.contentView addSubview:self.bubbleView];
         
         self.avatarImage = [[LYRUIAvatarImageView alloc] init];
-        [self.avatarImage setInitialViewBackgroundColor:LSLighGrayColor()];
+        self.avatarImage.backgroundColor = LSLighGrayColor();
         self.avatarImage.translatesAutoresizingMaskIntoConstraints = NO;
         [self.contentView addSubview:self.avatarImage];