+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 * `-[LYRUIMessageComposeTextView removeAttachements]` has been removed. It doesn't have a replacement since it was meant for internal use only.
 * The `placeHolderText` property of `LYRUIMessageComposeTextView` is now `placeholder`.
 * The unused `pendingBubbleViewColor` property of `LYRUIMessageCollectionViewCell` has been removed.
+* The `initialViewBackgroundColor` property of `LYRUIAvatarImageView` has been removed. Use `backgroundColor` instead.